后端：规范接口/external/user_out, 调整源码结构
