+++ backend/src/main/java/cn/alipapa/seat/bean/entity/Seat.java
@@ -1,4 +1,4 @@
-package cn.alipapa.seat.bean;
+package cn.alipapa.seat.bean.entity;
 
 public class Seat {
     private Integer id;
