+++ backend/src/main/java/cn/alipapa/seat/bean/entity/Serial.java
@@ -1,4 +1,4 @@
-package cn.alipapa.seat.bean;
+package cn.alipapa.seat.bean.entity;
 
 public class Serial {
     private Integer id;
