+++ backend/src/main/java/cn/alipapa/seat/bean/request/UserOutPOJO.java
@@ -1,32 +1,25 @@
-package cn.alipapa.seat.bean;
+package cn.alipapa.seat.bean.request;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 
 import java.util.Date;
 
-public class Record {
-    private Date enter_time;
+public class UserOutPOJO {
+    @JsonFormat(pattern = "HH:mm")
     private Date leave_time;
     private Integer id;
+    @JsonFormat(pattern = "yyyy-MM-dd")
     private Date record_date;
-    private String user_id;
+    private Integer user_id;
 
     @Override
     public String toString() {
-        return "Record{" +
-                "enter_time=" + enter_time +
-                ", leave_time=" + leave_time +
+        return "UserOutPOJO{" +
+                "leave_time=" + leave_time +
                 ", id=" + id +
                 ", record_date=" + record_date +
-                ", user_id='" + user_id + '\'' +
+                ", user_id=" + user_id +
                 '}';
-    }
-
-    public Date getEnter_time() {
-        return enter_time;
-    }
-
-    public void setEnter_time(Date enter_time) {
-        this.enter_time = enter_time;
     }
 
     public Date getLeave_time() {
@@ -53,11 +46,11 @@ public class Record {
         this.record_date = record_date;
     }
 
-    public String getUser_id() {
+    public Integer getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(String user_id) {
+    public void setUser_id(Integer user_id) {
         this.user_id = user_id;
     }
 }
